python-proj/pyproject.toml
[project]
name = "resize-screenshots"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pillow"]
